1.2.0 refactoring and improvements
The smtp and imap connector have been unified by the remote_controller
to have more similarity to the local_connector.

A awsome diff mode can, from now on, be used to display the false posities
and negatives, and if need be reassigning them to the corrent imap
folders.

On top of that the absencehandler has ben refactored to seperate the
main(prodctiv) function and the more development oriented debug mode.

Also the matcher now incooperates the entry validation into the parslib.